Fix Chromecast provider
